change field type

+++ src/main/java/cat/nyaa/infiniteinfernal/configs/WorldConfig.java
@@ -61,22 +61,22 @@ public class WorldConfig implements ISerializable {
     public Map<String, Double> trueDamage;
 
     @Serializable(name = "spawn.light.sky.min")
-    public byte minSkyLight = 0;
+    public int minSkyLight = 0;
 
     @Serializable(name = "spawn.light.sky.max")
-    public byte maxSkyLight = 0xf;
+    public int maxSkyLight = 0xf;
 
     @Serializable(name = "spawn.light.block.min")
-    public byte minBlockLight = 0;
+    public int minBlockLight = 0;
 
     @Serializable(name = "spawn.light.block.max")
-    public byte maxBlockLight = 0xf;
+    public int maxBlockLight = 0xf;
 
     @Serializable(name = "spawn.light.total.min")
-    public byte minLight = 0;
+    public int minLight = 0;
 
     @Serializable(name = "spawn.light.total.max")
-    public byte maxLight = 0xf;
+    public int maxLight = 0xf;
 
 
     private void initDefault() {
